text-extraction-function/requirements.txt
azure-functions
azure-cognitiveservices-vision-computervision
azure-storage-blob
python-dotenv
sendgrid
